fixed testing compilation issue
