+++ pkg/translator/appmesh/initialize.go
@@ -4,6 +4,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/aws/aws-sdk-go/service/appmesh"
 	gloov1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	"github.com/solo-io/go-utils/errors"
 	"github.com/solo-io/supergloo/pkg/api/custom/clients/kubernetes"
@@ -105,4 +106,58 @@ func getPodsForMesh(mesh *v1.Mesh, pods customkube.PodList) (AwsAppMeshPodInfo, 
 		}
 	}
 	return appMeshPods, appMeshPodList, nil
+}
+
+func upstreamForVNLabel(upstreams gloov1.UpstreamList, vnLabel, vnName string) (*gloov1.Upstream, error) {
+	for _, us := range upstreams {
+		labels := utils.GetLabelsForUpstream(us)
+		for key, value := range labels {
+			if key == vnLabel && value == vnName {
+				return us, nil
+			}
+		}
+	}
+	return nil, errors.Errorf("unable to find upstream with selector %s", vnLabel)
+}
+
+func podAppmeshConfig(info *podInfo, meshName, hostName string) (*appmesh.VirtualNodeData, *appmesh.VirtualServiceData) {
+	var vn *appmesh.VirtualNodeData
+	var vs *appmesh.VirtualServiceData
+	listeners := make([]*appmesh.Listener, len(info.ports))
+	for i, v := range info.ports {
+		port := int64(v)
+		protocol := "http"
+		listeners[i] = &appmesh.Listener{
+			PortMapping: &appmesh.PortMapping{
+				Protocol: &protocol,
+				Port:     &port,
+			},
+		}
+	}
+	vn = &appmesh.VirtualNodeData{
+		MeshName:        &meshName,
+		VirtualNodeName: &info.virtualNodeName,
+		Spec: &appmesh.VirtualNodeSpec{
+			Backends:  []*appmesh.Backend{},
+			Listeners: listeners,
+			ServiceDiscovery: &appmesh.ServiceDiscovery{
+				Dns: &appmesh.DnsServiceDiscovery{
+					Hostname: &hostName,
+				},
+			},
+		},
+	}
+
+	vs = &appmesh.VirtualServiceData{
+		MeshName:           &meshName,
+		VirtualServiceName: &hostName,
+		Spec: &appmesh.VirtualServiceSpec{
+			Provider: &appmesh.VirtualServiceProvider{
+				VirtualNode: &appmesh.VirtualNodeServiceProvider{
+					VirtualNodeName: &info.virtualNodeName,
+				},
+			},
+		},
+	}
+	return vn, vs
 }
